Add norwegian bokmål (nb_NO) locale support to google backend
When `norsk bokmål` language is selected from user preferences, the
locale code is set to `nb_NO`. The `SUPPORTED_LANG` already has entry
for for norwegian, but it is `no_NO`. If I understand correctly, this
locale code is deprecated. It is replaced by `nb_NO` for bokmål and
`nn_NO` for nynorsk. This commit only adds mapping from `nb_NO => 'no'`,
it does not remove the old `no_NO => 'no'` mapping.