FEATURE: Add translator error ProblemCheck for Microsoft
What does it add?
=================

This commit brings error messages from Microsoft Translator to the
administrator dashboard.

Specifically, a new InlineProblemCheck has been added, which reports the
error information in a human-readable way on the administrator dashboard
every time the translator server reports an error. When the translation
is successful, the error is considered to have been fixed and removed
from the dashboard.

ref: /t/134047