FIX: Allow admins to define locales that should be normalized
The criteria to allow a post to be translated is if the user's locale is different from the site's default locale.
For each translator (e.g. Google, Amazon, Azure), we maintain a list of mappings `SUPPORTED_LANG_MAPPING` for each locale. On top of that, th
is mapping is also used to define what language the post should be translated to. When using most translators available in the plugin, "en_GB" forums
 will translate text to "en" due to these mappings. This is not a problem for Discourse AI provider, since there are no restrictions on language.

This PR introduces a site setting `normalize_language_variants_map` so that admins can properly indicate if
A. A. They want `"pt" == "pt_BR" == "pt_PT"` or `"zh" == "zh_CN" == "zh_TW"` or
B. They would prefer that `"en"` is not the same as `"en_GB"`.