FEATURE: Show full topic translations (#205)
This feature shows a fully translated topic in the user's language. For now, the topic needs to already have translations for the translated topic to show. If there are posts that have not been translated, the original content will be shown.

The `translated_title` is returned via the `TopicViewSerializer` and displayed via the `applyTransformations` API.
The `translated_cooked` is returned via the `PostSerializer` and displayed via `decorateCooked`.

This takes a different approach from the previous PR https://github.com/discourse/discourse-translator/pull/199 which overrode the `fancy_title` and `cooked` directly.