FIX: Set a stripped post's locale to site default locale (#212)
In posts where there is a just an image, the stripped text that is sent for locale detection can be ...

```
<p></p>
```

... as we do not send HTML for the image for locale detection. The stripped text would result in a `""` returned for detected locale, which would result in 

```
Detected locale can't be blank
```

This commit defaults the post to the site's locale. This is needed so that we do not keep sending the post to check the locale.