FIX: Ensure old feature works with new and show translate button in correct scenarios (#215)
In the older implementation of translation, there exists these two site settings are part of the criteria that determines if 🌐 shows up per post

* `restrict translation by group` - Only allowed groups can translate
* `restrict translation by poster group` - Only allow translation of posts made by users in allowed groups. If empty, allow translations of posts from all users.

In a recent update with experimental features, we updated the cases where 🌐 will show up and caused a regression where the 🌐 shows up for everyone in `restrict translation by group`, ignoring `restrict translation by poster group`.

This commit makes sure these two site settings are adhered to when the experimental topic setting is turned off.