FIX: Fancify titles when using experimental topic translation (#224)
In the existing method of translation, we show the translated titles below the post cooked.

When using the same already-translated titles for the experimental feature, it would result in the H1 rendering into large text in the topic title.

We've been sending the `topic.title` in for translation, instead of the escaped `topic.fancy_title`, which works well for the existing method. I had considered changing our implementation to [send the fancy_titles (escaped) in for translation instead](https://github.com/discourse/discourse-translator/pull/222), but that would result in our existing methods titles being double-escaped.

To keep compatibility between existing and experimental, I will only escape it for the experimental feature. This commit does that.