DEV: Immediately send post for language detection for dual-text translations (#226)
Not too long ago, we moved language detections to a scheduled job. (see internal /t/134867/21)

But the new inline translation feature (when turned on) sends all content for translation the moment it is created. So far, we are not seeing an issue with load for this.

This commit reverts the batching for language detection to the original way, in-line with the new translation feature.

This commit also does a bit of cleanup - introducing three files to encapsulate the plugin event hooks into their own functionality (automatic translation, inline translation, dual-text translation)

This changes here will help to simplify the problem of 🌐 showing up in the post briefly, but in a subsequent commit.