UX: Remove flash of  🌐 when the post has yet to have its language determined (#229)
When a post is just created, we see the 🌐 appear for a brief moment, then disappear.

We typically want to show the 🌐 in case the post is not in the same language as the reader's. This works really well for old posts that do not have translation information. But since now all posts will be sent for locale detection on-create, it is a matter of seconds that the post's locale is determined.

This commit adds a buffer of 10 seconds that will prevent the 🌐 from appearing if 1. it was just created and 2. it doesn't have the locale yet.