DEV: Move saving translations outside of the API providers (#244)
Historically, each translation provider has been saving translation as part of doing the translation.

We are moving the saving outside of the providers, so `detect!` and `translate!` only does the detection and translation. Saving is now done in the base class so any future changes do not have to involve each translator.

This commit is to prepare / reduce complexity for the subsequent PR where we move to using raw to translate instead of cooked.