DEV: Translate raw content instead of cooked content (#245)
Raw is in general much smaller than cooked. Cooked tends to have a lot of html when it includes things like pulled oneboxes, mentions, images.

The gigantic cooked content also increases the amount of time it takes to translate content.

In sending raw, we can cook the translated content after it returns from the API. This also helps with sanitisation.