FIX: Revert translating raw for translator providers as they strip newlines (#250)
In https://github.com/discourse/discourse-translator/commit/6c197c561d3a2e725f722e9196e6cd7605f58f06, we changed to translate raw instead of cooked. This was largely due to DiscourseAI.

However some tests have shown that the original decision of translating cooked is likely due to translation providers like Google or Azure (and the rest) stripping newlines when translating. However these translator providers keep html as-is, so translating cooked is reasonable.

This commit reverts the other providers to use cooked, but keeps DiscourseAI as the only provider using raw.

The following is a trivial example, but in some more complex cases, the absence of certain characters (besides newline) removed by the provider when dealing with raw is a bit more jarring.