FIX: Use specific prompts for topic titles (#252)
This commit forks the existing DiscourseAi translator into a PostTranslator and TopicTranslator, with varying prompts to cater for the different content that can occur in post raw and topic titles.

We need to do this as titles tend to have a lot of text, and not have html or markdown, and the existing post content prompts are causing AI to pad the titles with html that does not exist.

This commit also improves the prompts for both types.