FEATURE: Add hidden settings to limit automatic translation scope (#255)
* DEV: Remove manual lock in favor of cluster_concurrency

This is a Discourse native feature that guarantees that we won't run
multiples of the same job at the same time.

* FEATURE: Add hidden settings to limit automatic translation scope

automatic_translation_backfill_max_age_days controls max age for
something to be eligible for translation

automatic_translation_backfill_limit_to_public_content limits the
automatic translation to posts and topics in categories who are public.