UX: Show extra details in failure popup when translation fails (#262)
When we translate a newly created topic (therefore topic title and post content), we send both title and cooked in separately. There is a chance that the title cannot be translated, but the post can be translated. The error does not make it obvious on which content the error is happening. 

This commit makes it precise to the reader if it is the "topic's title" or "post's content".