DEV: Clean up amazon translations due to parsing error (#305)
In https://github.com/discourse/discourse-translator/pull/203 we introduced a bug for the amazon provider, which was fixed in https://github.com/discourse/discourse-translator/pull/297.

This commit is a migration to purge the incorrect translations that have been stored. The translation will look like the following as the full response is saved instead of the specific attribute.

```
{translated_text: ...
```