DEV: Remove the anon language switcher and update settings (#308)
The anon language switcher feature is moving to core, so we are removing it from this plugin.

Similarly, in https://github.com/discourse/discourse/pull/32920, we introduced `experimental_content_localization_supported_locales`, but there was a borked migration that copies `automatic_translation_target_languages` (from this plugin) to `experimental_content_localization_supported_locales` (core) - the migration was a no-op. Language switcher relies on old setting, so it was good it was borked. This PR also adds a migration to do a rename as the old setting will not be used any more.

/t/154911